+++ src/main/java/cleytonorocha/com/github/back_dashboard/service/LocalService.java
@@ -1,41 +1,70 @@
 package cleytonorocha.com.github.back_dashboard.service;
 
-import java.util.List;
-import java.util.Optional;
-
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
+import cleytonorocha.com.github.back_dashboard.exception.ItemNotFoundException;
 import cleytonorocha.com.github.back_dashboard.model.entity.Local;
 import cleytonorocha.com.github.back_dashboard.model.repository.LocalRepository;
-import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.log4j.Log4j2;
 
+@Log4j2
 @Service
-@AllArgsConstructor
+@RequiredArgsConstructor
 public class LocalService {
 
     private final LocalRepository localRepository;
 
-    public List<Local> getAllLocals() {
-        return localRepository.findAll();
+    public Page<Local> findAll(Integer page, Integer linesPerPage, String orderBy, String direction) {
+        log.info("Fetching all locals with page: {}, linesPerPage: {}, orderBy: {}, direction: {}", page,
+                linesPerPage, orderBy, direction);
+
+        PageRequest pageRequest = PageRequest.of(page, linesPerPage, Sort.Direction.valueOf(direction), orderBy);
+        log.debug("PageRequest created: {}", pageRequest);
+
+        Page<Local> locals = localRepository.findAll(pageRequest);
+        log.debug("Locals fetched: {}", locals.getContent());
+
+        log.info("Returning {} locals", locals.getTotalElements());
+
+        return locals;
     }
 
-    public Optional<Local> getLocalById(Long id) {
-        return localRepository.findById(id);
+    public Local findById(Long id) {
+        log.info("Fetching local with id: {}", id);
+        return localRepository.findById(id)
+                .orElseThrow(() -> {
+                    log.warn("Local with id: {} not found", id);
+                    return new ItemNotFoundException();
+                });
     }
 
-    public Local createLocal(Local local) {
+    public Local save(Local local) {
+        log.info("Saving local: {}", local);
         return localRepository.save(local);
     }
 
-    public Local updateLocal(Long id, Local local) {
-        return localRepository.findById(id)
-                .map(map -> {
-                    map.setAddress(local.getAddress());
-                    return localRepository.save(map);
-                }).orElseThrow(() -> new RuntimeException("Local not found with id: " + id));
+    public Local update(Long id, Local local) {
+        log.info("Updating local with id: {}", id);
+        if (localRepository.existsById(id)) {
+            local.setId(id);
+            return localRepository.save(local);
+        } else {
+            log.warn("Local with id: {} not found", id);
+            throw new ItemNotFoundException();
+        }
     }
 
-    public void deleteLocal(Long id) {
-        localRepository.deleteById(id);
+    public void deleteById(Long id) {
+        log.info("Deleting local with id: {}", id);
+        if (localRepository.existsById(id)) {
+            localRepository.deleteById(id);
+        } else {
+            log.warn("Local with id: {} not found", id);
+            throw new ItemNotFoundException();
+        }
     }
 }
